gvpr: rewrite 'indexOf' to a simpler equivalent
This was initially attempted in 84efe93d5c5a4da0e778dbf01ed8550d2d774fdf and
then reverted in 51d787ece01fb4443db5eabd6932046b7b715dbc when it was discovered
this introduced a bug. The present commit fixes the, now obvious in retrospect,
bug with the original change of flipping the operands in the final subtraction.

This was validated by doing an exhaustive comparison of all strlen ≤2 inputs to
both the before and after function. Not bulletproof, but it is a strong signal
that the new version is functionally identical.