add new simple test_engines test
Upcoming commits in this series will make changes to gvFreeLayout and
we want to ensure that those changes don't introduce any memory leaks
when run with ASan, since gvFreeLayout calls layout-specific cleanup
functions.

This test is run in CI with ASan leak detection enabled. It currently
fails because of a memory leak in the osage layout.

Towards https://gitlab.com/graphviz/graphviz/-/issues/1800.