fix test suite 'dot' reliance on locale
This function was using the current system locale to encode and decode data sent
to Graphviz and received from Graphviz when using a textual output format. As a
result, encoding exceptions would occur if either the input or the output
contained non-ASCII characters and the system locale was not a UTF-8 one.
Apparently none of the current test suite hits this scenario. However, an
upcoming commit adds a test case that does.

This change forces the encoding and decoding to be done as UTF-8, which is also
what Graphviz unconditionally uses.