GTK plugin on_drawingarea1_scroll_event: squash -Wswitch-default
On GTK 2, the only valid scroll values are left, right, up, and down. GTK 3 adds
a fifth, smooth, capture scrolling in a combined x/y direction (diagonal). In
all cases except up and down, it seems clear this code intends to ignore the
action.

As of this commit, the GTK plugin builds warning-free on Linux.