Stable release 6.0.1
The 6.0.0 release was bungled, resulting in multiple commits¹ that claimed to be
Graphviz 6.0.0. ddd69c589483b31f589834e49ae902142cd4df38 of the website
repository² deleted the 6.0.0 download links and we have deleted the 6.0.0 tag
in this repository.

This commit updates the changelog to indicate there was no 6.0.0 release and the
next release will be 6.0.1. Hopefully things will go smoother this time around.

Unfortunately the situation with 6.0.0 was predictable. Discussion on #2202
covered a hole in the release procedure, which turned out to be the exact
problem that occurred.

¹ 9c4299601b40e5856df9a37cae78265ed2eb33b8 and
  091148d2910ac5f9b4d3b14578da2623f47e347b.
² https://gitlab.com/graphviz/graphviz.gitlab.io